+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -148,7 +148,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-11T00:35:33.650Z",
+    "2026-02-11T05:07:25.557Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {