+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-11T00:35:30.193Z",
+    "2026-02-11T05:07:22.360Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {